# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.8.0"
  },
  "paths": {
    "/agent/add_sub_recipes": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "add_sub_recipes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSubRecipesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Added sub recipes to agent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddSubRecipesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          }
        }
      }
    },
    "/agent/prompt": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "extend_prompt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Extended system prompt successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtendPromptResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          }
        }
      }
    },
    "/agent/resume": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "resume_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAgentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid working directory"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/session_config": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "update_session_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session config updated successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/start": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "start_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartAgentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid working directory"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/tools": {
      "get": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "get_tools",
        "parameters": [
          {
            "name": "extension_name",
            "in": "query",
            "description": "Optional extension name to filter tools",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "description": "Required session ID to scope tools to a specific session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tools retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/update_provider": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "update_agent_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Provider updated successfully"
          },
          "400": {
            "description": "Bad request - missing or invalid parameters"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/update_router_tool_selector": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "update_router_tool_selector",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouterToolSelectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tool selection strategy updated successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddSubRecipesRequest": {
        "type": "object",
        "required": [
          "sub_recipes",
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "sub_recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubRecipe"
            }
          }
        }
      },
      "AddSubRecipesResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ExtendPromptRequest": {
        "type": "object",
        "required": [
          "extension",
          "session_id"
        ],
        "properties": {
          "extension": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "ExtendPromptResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ExtensionData": {
        "type": "object",
        "description": "Extension data containing all extension states\nKeys are in format \"extension_name.version\" (e.g., \"todo.v0\")",
        "additionalProperties": {}
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "MessageMetadata": {
        "type": "object",
        "description": "Metadata for message visibility",
        "properties": {
          "agentVisible": {
            "type": "boolean",
            "description": "Whether the message should be included in the agent's context window"
          },
          "userVisible": {
            "type": "boolean",
            "description": "Whether the message should be visible to the user in the UI"
          }
        }
      },
...
      "PermissionLevel": {
        "type": "string",
        "description": "Enum representing the possible permission levels for a tool.",
        "enum": [
          "always_allow",
          "ask_before",
          "never_allow"
        ]
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "json_schema": {
            "nullable": true
          }
        }
      },
      "ResumeAgentRequest": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "SessionConfigRequest": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Response"
              }
            ],
            "nullable": true
          },
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "SessionMetadata": {
        "type": "object",
        "description": "Metadata for a session, stored as the first line in the session file",
        "required": [
          "working_dir",
          "description",
          "message_count"
        ],
        "properties": {
          "accumulated_input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of input tokens used in the session. Accumulated across all messages.",
            "nullable": true
          },
          "accumulated_output_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of output tokens used in the session. Accumulated across all messages.",
            "nullable": true
          },
          "accumulated_total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used in the session. Accumulated across all messages (useful for tracking cost over an entire session).",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "A short description of the session, typically 3 words or less"
          },
          "extension_data": {
            "$ref": "#/components/schemas/ExtensionData"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of input tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "description": "Number of messages in the session",
            "minimum": 0
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of output tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "description": "ID of the schedule that triggered this session, if any",
            "nullable": true
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used in the session. Retrieved from the provider's last usage.",
            "nullable": true
          },
          "working_dir": {
            "type": "string",
            "description": "Working directory for the session",
            "example": "/home/user/sessions/session1"
          }
        }
      },
...
      "StartAgentRequest": {
        "type": "object",
        "required": [
          "working_dir"
        ],
        "properties": {
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "working_dir": {
            "type": "string"
          }
        }
      },
      "StartAgentResponse": {
        "type": "object",
        "required": [
          "session_id",
          "metadata",
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/SessionMetadata"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "SubRecipe": {
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "sequential_when_repeated": {
            "type": "boolean"
          },
          "values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
...
      "ToolInfo": {
        "type": "object",
        "description": "Information about the tool used for building prompts",
        "required": [
          "name",
          "description",
          "parameters"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parameters": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permission": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PermissionLevel"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "UpdateProviderRequest": {
        "type": "object",
        "required": [
          "provider",
          "session_id"
        ],
        "properties": {
          "model": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "UpdateRouterToolSelectorRequest": {
        "type": "object",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
...
    }
  }
}